% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGQ.R
\name{EGQ}
\alias{EGQ}
\title{Expected Genotype Quality}
\usage{
EGQ(DP, error, prior)
}
\arguments{
\item{DP}{read depth}

\item{error}{allelic error}

\item{prior}{numeric vector of length ploidy + 1}
}
\value{
numeric vector of length ploidy + 1
}
\description{
Expected Genotype Quality for Binomial Model
}
\details{
Expected GQ conditional on the true allele dosage (0,1,2,...ploidy)
}
