% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_cM.R
\name{interpolate_cM}
\alias{interpolate_cM}
\title{Interpolate genetic distances}
\usage{
interpolate_cM(map, df = 8, max.extend = 5)
}
\arguments{
\item{map}{Data frame with four columns: marker, chrom, bp, cM}

\item{df}{Degrees of freedom for the spline}

\item{max.extend}{Maximum distance (in cM) to extrapolate beyond the end of the input map}
}
\value{
List containing
\describe{
\item{map}{Map data frame with additional column named cM.spline}
\item{plot}{ggplot2 object}
}
}
\description{
Creates monotone spline between physical (bp) and genetic (cM) distance
}
\details{
The input map can be generated by merging an existing genetic map with positions in bp and cM with additional markers with only bp information. Interpolation is based on minimizing the mean-squared error between the original and interpolated positions in cM. For the \code{df} and \code{max.extend} parameters, use either a single integer (if same for all chromosomes) or a vector of length equal to the number of chromosomes.
}
