% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_map_fn.R
\name{inverse_map_fn}
\alias{inverse_map_fn}
\title{Inverse map function}
\usage{
inverse_map_fn(x, model)
}
\arguments{
\item{x}{map distance (cM)}

\item{model}{Either "Haldane" or "Kosambi"}
}
\value{
recombination frequency
}
\description{
Computes recombination frequency from map distance
}
