% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDbin.R
\name{LDbin}
\alias{LDbin}
\title{Create marker bins based on LD}
\usage{
LDbin(geno, r2.thresh = 0.99)
}
\arguments{
\item{geno}{matrix of haplotype dosages (markers x indiv)}

\item{r2.thresh}{threshold for binning}
}
\value{
List containing
\describe{
\item{bins}{data frame with two columns: marker,bin}
\item{geno}{genotype matrix for the bins}
\item{r2}{r2 matrix for the bins}
}
}
\description{
Create marker bins based on LD
}
\details{
Bins are created based on hierarchical clustering with \code{hclust} and \code{method='single'}, using \eqn{1-r^2} as the dissimilarity metric. The argument \code{r2.thresh} controls the height for cutting the dendrogram to create the bins. The marker with the least missing data for each bin is chosen to represent it.
}
