% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG.R
\name{LG}
\alias{LG}
\title{Make linkage groups based on clustering}
\usage{
LG(LODmat, thresh = seq(2, 20, by = 2))
}
\arguments{
\item{LODmat}{matrix of LOD scores for the marker bins}

\item{thresh}{numeric vector of thresholds for clusterings}
}
\value{
Either a ggplot2 object or data frame of linkage groups (see Details)
}
\description{
Make linkage groups based on clustering
}
\details{
If \code{thresh} is a numeric vector with multiple LOD thresholds, the function returns a plot showing the number of markers per LG. If \code{thresh} is a single value, the function returns a data frame with the LG assignment for each marker. LGs are numbered from the largest to smallest group.
}
