% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fn.R
\name{map_fn}
\alias{map_fn}
\title{Map functions}
\usage{
map_fn(r, model)
}
\arguments{
\item{r}{recombination frequency}

\item{model}{Either "Haldane" or "Kosambi"}
}
\value{
Map distance in cM
}
\description{
Computes cM map distance from recombination frequency
}
