% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_markers.R
\name{order_markers}
\alias{order_markers}
\title{Order markers by solving the TSP}
\usage{
order_markers(x)
}
\arguments{
\item{x}{distance matrix}
}
\value{
a list containing
\describe{
\item{order}{optimized order as a vector of integers}
\item{distance}{sum of adjacent distances}
}
}
\description{
Order markers by solving the TSP
}
\details{
Uses R package seriation to minimize the distance between adjacent markers. For example, \code{x} could be a matrix of recombination frequencies or monotone decreasing transformation of LOD scores.
}
