% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coverage.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot marker coverage of the genome}
\usage{
plot_coverage(map, limits = NULL)
}
\arguments{
\item{map}{data frame with columns chrom & position}

\item{limits}{optional data frame with columns chrom & position, with the maximum length for each chromosome}
}
\value{
ggplot2 variable
}
\description{
Plot marker coverage of the genome
}
\details{
If \code{limits} not provided, then the maximum values in \code{map} are used.
}
