% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geno.R
\name{plot_geno}
\alias{plot_geno}
\title{Graphical genotyping}
\usage{
plot_geno(geno, map = NULL)
}
\arguments{
\item{geno}{genotype matrix (markers x indiv)}

\item{map}{optional data frame with 3 columns (marker, chrom, position)}
}
\value{
ggplot object
}
\description{
Graphical genotyping
}
\details{
Input matrix \code{geno} should have rownames attribute that matches marker names in the first column of \code{map} (when present).
}
