% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genofreq.R
\name{plot_genofreq}
\alias{plot_genofreq}
\title{Plot and filter markers based on genotype frequency vs position}
\usage{
plot_genofreq(geno, thresh = 0.1, span = 0.3)
}
\arguments{
\item{geno}{haplotype dosage matrix (markers x indiv)}

\item{thresh}{threshold for removing markers (see Details)}

\item{span}{parameter to control degree of smoothing for spline (higher = less smooth)}
}
\value{
List containing
\describe{
\item{outliers}{character vector of marker names}
\item{plot}{ggplot2 variable}
}
}
\description{
Plot and filter markers based on genotype frequency vs position
}
\details{
Genotypes should be coded 0,1,2. Markers are removed if their residual to the fitted spline exceeds \code{thresh}. Markers are assumed to be ordered. Function designed to be used for one chromosome.
}
