% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genoprob.R
\name{plot_genoprob}
\alias{plot_genoprob}
\title{Plot genotype probabilities for one chromosome}
\usage{
plot_genoprob(genoprob, map)
}
\arguments{
\item{genoprob}{matrix (markers x genotypes) of probabilities for one individual}

\item{map}{map data frame (markers,chrom,position)}
}
\value{
ggplot object
}
\description{
Plot genotype probabilities for one chromosome
}
\details{
Names for the genotypes are taken from the colnames of \code{genoprob}.
}
