% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_haplofreq.R
\name{plot_haplofreq}
\alias{plot_haplofreq}
\title{Plot and filter markers based on haplotype frequency vs position}
\usage{
plot_haplofreq(geno, thresh = 0.1, span = 0.3)
}
\arguments{
\item{geno}{haplotype dosage matrix (markers x indiv)}

\item{thresh}{threshold for removing markers (see Details)}

\item{span}{parameter to control degree of smoothing for spline (higher = less smooth)}
}
\value{
List containing
\describe{
\item{outliers}{index of outlier markers}
\item{plot}{ggplot2 variable}
}
}
\description{
Plot and filter markers based on haplotype frequency vs position
}
\details{
This functions works for mapping populations derived from two inbred founders. The \code{geno} matrix must be coded based on the dosage (0,1,2) of one founder haplotype.

Markers are removed if their residual to the fitted spline exceeds \code{thresh}. Markers are assumed to be ordered. Function designed to be used for one chromosome.
}
