% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_LD.R
\name{plot_LD}
\alias{plot_LD}
\title{Plot LD vs distance}
\usage{
plot_LD(r2, map, max.pair = 10000, dof = 5)
}
\arguments{
\item{r2}{squared correlation matrix}

\item{map}{data frame with 3 columns (marker, chrom, position)}

\item{max.pair}{maximum number of r2 pairs for the spline}

\item{dof}{degrees of freedom for the spline}
}
\value{
ggplot object
}
\description{
Plot LD vs distance
}
\details{
A monotone decreasing, convex spline is fit using R package \code{scam}. The input matrix \code{r2} should have rownames attribute that matches marker names in the first column of \code{map}.
}
