% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_r2.R
\name{plot_r2}
\alias{plot_r2}
\title{Plot heatmap of LD}
\usage{
plot_r2(r2)
}
\arguments{
\item{r2}{squared correlation matrix}
}
\value{
ggplot2 variable
}
\description{
Plot heatmap of LD (squared correlation)
}
