% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_square.R
\name{plot_square}
\alias{plot_square}
\title{Plot square (dis)similarity matrix}
\usage{
plot_square(data, lims = NULL)
}
\arguments{
\item{data}{square matrix}

\item{lims}{numeric 3-vector with the low,mid,high points for the colors}
}
\value{
ggplot2 variable
}
\description{
Plot square (dis)similarity matrix
}
\details{
Can be used to plot squared correlation, recomb frequency, LOD and more. By default, \code{lims} equals (0,median,max)
}
