% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbit_diallel.R
\name{rabbit_diallel}
\alias{rabbit_diallel}
\title{Make RABBIT input files for diploid diallel population}
\usage{
rabbit_diallel(ped, geno, geno.founder, map, outstem)
}
\arguments{
\item{ped}{data frame with pedigree (pop,parent1,parent2)}

\item{geno}{list of genotype matrices (markers x indiv), one for each population in \code{ped}}

\item{geno.founder}{matrix of genotype data for the founders (markers x indiv)}

\item{map}{genetic map (marker,chromosome,position)}

\item{outstem}{name for output files}
}
\description{
Make RABBIT input files for diploid diallel population
}
\details{
Populations must be numbered in \code{ped} corresponding to their position in \code{geno}. Founders are not included in \code{ped}. All genotype matrices must have identical markers. Genetic map position should be in cM. Genotypes need to be coded according to RABBIT format.
}
