% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbit_read.R
\name{rabbit_read}
\alias{rabbit_read}
\title{Parse output from RABBIT MagicReconstruct}
\usage{
rabbit_read(rabbit.file, ML.file = NULL, diaQTL.file = NULL)
}
\arguments{
\item{rabbit.file}{name of RABBIT output file}

\item{ML.file}{name of most likely genotype file to create}

\item{diaQTL.file}{name of diaQTL genotype file to create}
}
\value{
data frame defining the genotypes
}
\description{
Parse output from RABBIT MagicReconstruct
}
\details{
Two different file formats can be created. The \code{ML.file} contains the most likely (i.e., posterior maximum) genotype for each individual at each marker. The \code{diaQTL.file} contains the full distribution of genotype probabilities in the format required by the diaQTL R package (\code{diaQTL.file}). The default value for each filename is NULL, which generates no file.
}
