% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S1_haplo.R
\name{S1_haplo}
\alias{S1_haplo}
\title{Phase S1 parent and reconstruct progeny in terms of parental haplotypes}
\usage{
S1_haplo(geno, error, n.core = 1)
}
\arguments{
\item{geno}{ordered genotype matrix (markers x indiv) for one chromosome}

\item{error}{average genotype error to use for the HMM}

\item{n.core}{number of cores (default is 1)}
}
\value{
List containing
\describe{
\item{parent}{two column matrix (rows = markers) with the haplotypes for the parent}
\item{progeny}{matrix with progeny reconstructed based on dosage of the second parental haplotype}
}
}
\description{
Phase S1 parent and reconstruct progeny in terms of parental haplotypes
}
\details{
It is assumed that only segregating markers are present. Progeny reconstruction occurs using an HMM with a uniform transition probability matrix, based on an average recombination frequency \code{r}, and a uniform model for the genotype \code{error}.
}
