% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S1_selection.R
\name{S1_selection}
\alias{S1_selection}
\title{Signatures of selection in S1 populations}
\usage{
S1_selection(x)
}
\arguments{
\item{x}{Vector of progeny counts for the three possible genotypes (AA, AB, BB)}
}
\value{
Data frame with columns "s" for the ML selection coefficient and "score" for the -log10(p) value.
}
\description{
Signatures of selection in S1 populations
}
\details{
The null hypothesis is no selection, in which case the expected frequency of genotypes is (AA = 1/4, AB = 1/2, BB = 1/4). Four alternate hypotheses are tested, corresponding to 1. gametic selection in one sex, 2. gametic selection in both sexes, 3. zygotic selection against one homozygote, 4. zygotic selection against both homozygotes. The selection coefficient (s) is estimated by maximum likelihood on the interval [-1,1], where positive values correspond to selection against A or AA and negative values are for selection against B or BB. For model 4, positive values represent selection against homozygotes, while negative values indicate selection against heterozygotes.  P-values are computed based on the likelihood ratio test; in other words, the change in deviance is assumed to be chi-squared (df = 1) distributed under the null hypothesis.
}
