% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGQ.R
\name{EGQ}
\alias{EGQ}
\title{Expected Genotype Quality}
\usage{
EGQ(depth, error, ploidy, prior)
}
\arguments{
\item{depth}{read count}

\item{error}{allelic error}

\item{ploidy}{ploidy}

\item{prior}{numeric vector of length ploidy + 1}
}
\value{
numeric scalar
}
\description{
Expected Genotype Quality for Binomial Model
}
\details{
As defined in Matias et al. (2019), EGQ is the PHRED-scaled expected error of the genotype call, conditional on the true genotype. This function returns EGQ for the genotype most frequently miscalled, which is the balanced heterozygote (i.e., ploidy/2).
}
\references{
Matias et al. (2019) Plant Genome 12:190002. https://doi.org/10.3835/plantgenome2019.01.0002
}
