% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_map.R
\name{genetic_map}
\alias{genetic_map}
\title{Multi-point estimation of a genetic map}
\usage{
genetic_map(x, n.point = 3)
}
\arguments{
\item{x}{matrix of pairwise map distances (cM) between the marker-bins for one chromosome}

\item{n.point}{Number of points used for estimation}
}
\value{
data frame with columns marker,position (in cM)
}
\description{
Multi-point estimation of a genetic map
}
\details{
Argument \code{n.point} controls how many pairwise distances are used in the linear regression. \code{n.point=2} means only adjacent bins; \code{n.point=3}  means adjacent bins and bins with one intervening marker, etc. Marker names taken from the rownames attribute of \code{x}. For multi-point estimation to be effective, \code{x} should be based on marker-bins.
}
