% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDbin.R
\name{LDbin}
\alias{LDbin}
\title{Create marker bins based on LD}
\usage{
LDbin(geno, r2.thresh)
}
\arguments{
\item{geno}{matrix of haplotype dosages (markers x indiv)}

\item{r2.thresh}{threshold for binning}
}
\value{
List containing
\describe{
\item{bins}{list of character vectors, containing the markers for each bin}
\item{r2}{matrix of squared correlations between bins}
}
}
\description{
Create marker bins based on LD
}
\details{
Bins are created based on hierarchical clustering with \code{hclust} and \code{method='single'}, using \eqn{1-r^2} as the dissimilarity metric. The argument \code{r2.thresh} controls the height for cutting the dendrogram to create the bins. The marker chosen to represent each bin is the one with the least missing data.
}
