% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{LL}
\alias{LL}
\title{Log-likelihood for inbred line-derived mapping populations}
\usage{
LL(r, counts, pop.type)
}
\arguments{
\item{r}{recombination frequency}

\item{counts}{3x3 contingency table for haplotype dosages 0,1,2}

\item{pop.type}{One of the following: "DH","BC","F2"}
}
\value{
log-likelihood
}
\description{
Log-likelihood for inbred line-derived mapping populations
}
\details{
The argument \code{counts} can be constructed using the \code{table} function for two markers. Genotype coding must represent dosage of a founder haplotype. For BC populations, possible allele dosages are {0,1}. For DH pops, it is {0,2}. For F2 pops, it is {0,1,2}.
}
