% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEL.R
\name{MLEL}
\alias{MLEL}
\title{Max Likelihood Estimation of Linkage}
\usage{
MLEL(geno, pop.type, LOD, n.core = 1)
}
\arguments{
\item{geno}{Matrix of haplotype dosages (markers x indiv)}

\item{pop.type}{One of the following: "DH","BC","F2"}

\item{LOD}{Logical, whether to return LOD (TRUE) or recomb freq (FALSE)}

\item{n.core}{For parallel execution on multiple cores}
}
\value{
Matrix with RF or LOD
}
\description{
Max Likelihood Estimation of Linkage
}
\details{
Can be used to estimate either the LOD score or recombination frequency, depending on the value of \code{LOD}. Genotype coding must represent dosage of a founder haplotype. For BC populations, possible allele dosages are 0,1. For DH and RIL pops, it is 0,2. For F2 pops, it is 0,1,2.
}
