% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_haplo.R
\name{plot_haplo}
\alias{plot_haplo}
\title{Graphical genotyping}
\usage{
plot_haplo(geno, map)
}
\arguments{
\item{geno}{genotype matrix (markers x indiv)}

\item{map}{data frame with 3 columns (marker, chrom, position)}
}
\value{
ggplot object
}
\description{
Graphical genotyping
}
\details{
Input matrix \code{geno} should have rownames attribute that matches marker names in the first column of \code{map}.
}
