% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plots data against map}
\usage{
plot_map(data)
}
\arguments{
\item{data}{data frame with 3 columns: chrom, position, y (the plotting variable)}
}
\value{
ggplot
}
\description{
Plots data against map
}
