% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbit_read.R
\name{rabbit_read}
\alias{rabbit_read}
\title{Parse output from RABBIT MagicReconstruct}
\usage{
rabbit_read(filename, outfile1 = "MLgeno.csv", outfile2 = NULL)
}
\arguments{
\item{filename}{name of RABBIT output file}

\item{outfile1}{name of ML genotype file to create}

\item{outfile2}{name of diaQTL genotype file to create}
}
\description{
Parse output from RABBIT MagicReconstruct
}
\details{
Generates output files with the maximum likelihood (ML) genotype (\code{out1file}) and genotype probabilities for the diaQTL package (\code{out2file}).
}
