% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S1_selection.R
\name{S1_selection}
\alias{S1_selection}
\title{Signatures of selection in S1 populations}
\usage{
S1_selection(data, alpha = 0.05)
}
\arguments{
\item{data}{data frame with columns: marker, chrom, position, AA, AB, BB. Columns 4-6 have count data.}

\item{alpha}{significance level}
}
\value{
list with "plot" and "table" of results:
\describe{
\item{model}{name of best model}
\item{s}{selection coefficient}
\item{score}{-log10(p) value}
}
}
\description{
Signatures of selection in S1 populations
}
\details{
Genotypes must be coded based on the S1 parental haplotypes, not markers.

The null hypothesis is no selection, in which case the expected frequency of genotypes is (AA = 1/4, AB = 1/2, BB = 1/4). Two alternate hypotheses are tested for gametic selection: 1.selection in one sex, 2.selection in both sexes. Two models of zygotic selection are also tested: 1.selection against one homozygote, 2.selection against both homozygotes. The selection coefficient equals the sum of the absolute differences between the observed and expected frequencies. Positive values correspond to selection against A or AA, negative values for selection against B or BB. For zygotic2, positive (negative) values represent selection against (for) homozygotes. 

P-values are computed based on the likelihood ratio test; in other words, the change in deviance is assumed to be chi-squared distributed under the null hypothesis.
}
